Move Matrix serialization to separate file and remove spurious headers
